bit more rearranging in statusManager/propertyManager; cleaner connection with viewer. I'll post a diagram showing the connections in http://www.stolaf.edu/people/hansonr/jmol/docs/dev prototype is at http://www.stolaf.edu/people/hansonr/jmol/test/json/callback.htm and http://www.stolaf.edu/people/hansonr/jmol/test/json/json.htm
Relevant changes include the change of name of the function from getProperty() to getProperty(), getPropertyAsString(), getPropertyAsJSON() -- all of which work for any property, so you can choose. This is a generalization that could be used by other apps, because nothing (other than applet-specific callback calls) is situated in the applet-specific code anymore.