OK, two issues here:
1) I think I fixed a bug in Dots.java -- Miguel, take a look, but there was a null pointer problem in Dots.java, where colixesConvex[i] didn't exist because colixesConvex had never been initialized. Something about (mad) -- this must be yours or egon's work, I am guessing. I "fixed" it by adding:

    if ("colorConvex" == propertyName) {
      if(colixesConvex == null)
        return;

but I would appreciate you looking and seeing if there is something else that really should be done rather than this.

2) The second fix is a PROPOSED fix for a very tricky problem: We need to be able to introduce new commands, including new tokens. At the same time, users need to be assured that what they do will most likely run with future versions. It's an old problem, of course, but this has to do with "define". Currently a user cannot redefine anything but "predefined sets". OK, that's reasonable. But when I cleaned up "frame" I made a couple more tokens, specifically, "prev" for "previous frame". Now, it turns out that Rolf has a very complex set of defines in his application, and one of them happens to be "prev". Of course, adding "prev" as a token for frame breaks this page.

Here's my solution: Specifically in 'define' commands, we DO let the user redefine terms, but we throw up a clear warning as to possible effects -- some commands (in this case frame) may not work properly after this definition. I believe that users sophisticated enough to use 'define' are sophisticated enough to look at the console and consider these warnings. But this way it doesn't break the page.

It was a simple addition. To undo this implementation, just look for where  'isDefine' is made true in Compiler.java and comment out that line. But I'd like us to think about this more if necessary before undoing this change.

Alternatively (and I guess I'm ok with this), perhaps we could agree: absolutely no more tokens. All future additions must be string matches, but not tokens. If so, we need to clearly state that in Tokens.java. And we have to realize that it hinders future development.


Bob