working on visibility; dissociating clickability from visibility (ahem, as Miguel pointed out would be necessary initially)
(1) changed Atom.visibilityFlags to Atom.clickabilityFlags
(2) some experimentation with keeping a running total of the
number of displayed bonds to an atom was NOT successful; still in there
though. backbone visibility same deal. I'll clean this up.

Change here is that if any cartoon or trace or such is visible,
then all atoms inherit group visibility. Just testing this.