adds getProperty("stateInfo"), which combines:
(a) a differential listing of atom and bond information
(b) modelInfo
(c) animationInfo
(d) polymerInfo
(e) shapeInfo

this should go a long way toward generating a "state" of
the model. All the atom and bond sizes and colors are there;
the shapeVisibilityFlags indicate what sort of shape is enabled; 
This isn't perfect by any means. Still no mads for the special
shapes. No labels. But it's all I have time for currently. 

The differential idea is not obvious. 

See ModelManager.getAllStateInfo().