OK, a new idea for applet synchronization. test is at http://www.stolaf.edu/people/hansonr/jmol/test/json/sync.htm
It's not fully articulated, but the basic idea being tested
here is direct applet-applet communication via a static appletRegistry

Implementation, using Jmol-new.js:

function jmolSync(mode, targetSuffix) {
    _jmolCheckBrowser();
    if (targetSuffix == "all") {
      with (_jmol) {
	for (var i = 0; i < appletSuffixes.length; ++i) {
	  var applet = _jmolGetApplet(appletSuffixes[i]);
          if (applet) applet.syncScript(mode);
        }
      }
    } else {
      var applet=_jmolGetApplet(targetSuffix);
      if (applet) applet.syncScript(mode);
    }
}

where <mode> is one of:

"ON"   --driver of synchronizations
"SYNC" --turn on and sync, but not as a driver
"OFF"  --not synced (not quite working yet)

This uses a new public applet method, syncScript(), which
allows control of synchronization and passing of a reorientation script.

The method uses NO JavaScript -- the applets identify themselves
within a common registry and then talk directly to each other.

Part of the idea here is to experiment with this sort of direct
communication -- calling functions in a different applet.
I suspect there will be browser/platform issues.

Just a start. All I have time for today.