(1) in new draw/spin/rotate/center changes to {braces} for coordinates and $name for draw object references; compiler changed as necessary for this; might have some clean up to do, but at least it's out of the Rasmol parser.
This is a sounder strategy. 

(2) bug found (that I introduced, sorry!) in measurements not
rendering properly. This was due to my forgetting that torsions 
can have negative values.

(3) free to revert if you like, but what I learned today is that 
Float.MIN_VALUE is a POSITIVE number, not "negative infinity". 
I found some places where changing to -Float.MAX_VALUE was 
appropriate and did so. If I'm wrong in isosurface and mesh (that you 
really want MIN_VALUE "1E-45" there, not what I've changed that to, 
-MAX_VALUE "-1E38", let me know.