Fixed reading of two additional Gaussian files.
I had to increase the number of lines that the Resolver uses to try to determine what file type it is for the ch2chfme_reagent.out file.
If that is a bad thing to do, you can set it back to 4, but that is not sufficient to determine the file type for a file like this one.