polyhedra: two tests only
(1) adds optional # of bonds to force -- 4 or 6 for standard
"BONDS" option:

 polyhedra BONDS 4
 polyhedra BONDS 6

(2) adds "NBONDS" same as "BONDS" but allows 
    any number of bond connections as:

 polyhedra NBONDS 
 polyhedra NBONDS 5
 polyhedra NBONDS 7

etc.

(3) adds "FACETS" as an option instead of "BONDS" 

 polyhedra FACETS
 polyhedra FACETS n   where n is any number, not just 4 or 6.

 -- displays each "face" as a set of three (or four) "facets" 
(triangles connecting the vertices to the central atom, not 
a single vertex-to-vertex-to-vertex face. This allows for 
seeing "into" the model and still observe the central atom. 

demo will be at http://www.stolaf.edu/people/hansonr/jmol/test/json/poly.htm