polyhedra adds
polyhedra maxFactor <x> facets
polyhedra facetCenterOffset <x> facets

maxFactor is the factor that is used to determine whether the
distance between vertices relative to the central bond length
is short enough to define a face. Default is 1.85, but this can be changed. 

facetCenterOffset sets the distance from the deepest point of 
of the facet to the center atom. Default is 0.25 Angstroms, but
this can be changed to give other sometimes MOST interesting
effects.