polyhedrons finished, with modifications;
"facets" --> "collapsed"
"bonds" --> new method
"radius" --> new method

Essentially, a total rewrite of Polyhedra.java


polyhedra 

Jmol supports polyhedral representation of molecular structures, including
tetrahedra, octahedra, and many other geometric forms. 

Centers. Potential polyhedra centers are defined either by the currently selected set of 
atoms or by explicitly indicating an atom expression within the polyhedra command. 

polyhedra (chlorine)
polyhedra (atomno=1) 2.0
polyhedra (atomno=1 or atomno=3) 4
select (atomno < 3); polyhedra 4

Vertices. From this set of centers, polyhedra can be formed based upon a 
specified number of bonds (e.g., polyhedra 6) or upon a maximum distance expressed
as a decimal number (e.g., polyhedra 2.0). In addition, you can add the keyword "TO" 
and a second atom expression to indicate that only atoms matching this expression 
should be selected for vertices. The default is to use bonds, 
but if a radius and a number of bonds is indicate, then the
number of "bonds" simply means the number of vertices within
this radius, actually "bonded" or not.

polyhedra (silicon) TO (oxygen) 4
polyhedra (zinc) TO (oxygen or nitrogen) 5.0

Options. Several additional options may be used to adjust the appearance of the polygons. 
These include:

COLLAPSED

Collapse the polygon into the center atom, as though all the volume had been removed, leaving
only a "skin." This allows visualization of the central atom. 

EDGES


Highlight the edges of the polyhedra.

FRONTEDGES

Highlight only the front edges of the polyhedra (Useful when color is translucent.)

MAXFACTOR [decimal number] (default 1.85)

Determines the cutoff point used to determine faces. Vertex-vertex distances longer than
this factor times the average center-vertex distance are discarded. In certain circumstances
this parameter can be adjusted to produce better polygons.

FACECENTEROFFSET [decimal number] (default 0.25)

This absolute offset, in Angstroms, sets the position of the point at the base of the 
cavity in a collapsed polygon. The value 0.0 indicates that the base of the cavity should
be the central atom and is not recommended. Adjusting this parameter can lead to interesting
effects.

Additional commands that can be invoked after polygons have been created include 
the following. Note that each can include a specified atom set or use the current 
selection so as to operate only on specific polyhedra based on their center atoms.

polyhedra ON/OFF{default: ON}

    Show or hide polyhedra whose central atom is selected.

polyhedra DELETE

    Delete the polyhedra associated with central atoms that are selected.

polyhedra EDGES

    As above

polyhedra FRONTEDGES

    As above

polyhedra NOEDGES

    Turns off edges