Fixes improper handling of missing sequence number in mmCIF files.
1d66.cif:

HETATM 1711 CD CD    . CD  H 4 .  ? 33.200 64.497 45.835 1.00 39.60 ? ? ? ? ? 
42  CD  ? CD  1 
HETATM 1712 O  O     . HOH I 5 .  ? 40.594 60.277 53.968 1.00 16.15 ? ? ? ? ? 
301 HOH ? O   1 

Note the "?" in that last line. 
That "?" is supposed to be the group sequence number, 
42 in the case of CD, 301 in the case of HOH. 
The program that generates these CIF files is ignoring 
the group sequence number for HETATM records. 
 
Now consider the PDB file:

1d66.pdb:

HETATM 1715 CD    CD    42      33.200  64.497  45.835  1.00 39.60      1D661802
HETATM 1716  O   HOH   301      40.594  60.277  53.968  1.00 16.15      1D661803

Jmol could not distinguish beteen the two numbers because 
they are not there in the mmCIF file.
So it was not considering HOH to be different from CD, 
and was assigning it that group name. Instead of a group 
sequence number, Jmol was assigning to both CD and HOH "Integer.MIN_VALUE".

My fix assigns a unique negative number to atom.sequenceNumber and
also collects atomSerial, only for informational purposes at this point. 

Bob Hanson