Adds support for Fujitsu CAChe Chemical Structure File (CSF) format. CAChe is one of the most popular visualization/calculation tools at the undergraduate level, particularly among Macintosh users.
CsfReader was modeled after CifReader, as both formats involve a
flexible-field record format with variable but predefined column-based fields.

Vibration information is turned into an initial static frame followed
by a frame for each normal mode. We may want to look at the default
scaling for the vibrational vector magnitudes, but I don't know that
we have any standard for that among file types, anyway. 

Required for this was the addition of AtomSetCollection.cloneFirstAtomSetWithBonds(int)
I know passing the number of bonds is not optimal, but I couldn't find the
analogy to atomSetAtomCounts[0] in terms of bonds. So that should be looked at.
(CsfReader should not be responsible for tracking how many bonds have been created.)

Nonetheless, this will be a welcome addition, I think.

Resolver just looks for a line STARTING with "local_transform". 
A search of the entire Jmol example data file set indicates no
other file types contain this token.

Bob Hanson