10.3 bug fix #1: CifReader applying strand mapping to atoms but not structures. Issue found with 1pgb.cif not showing secondary structure is fixed.
But there is an interesting difference between cif and pdb that
perhaps will not be able to be fixed. This has to do with the 
ORDER with which helix/turn/strand are encountered in the file.
What to do if, as in this case, residues are listed twice, once
as strand and once as turn?

Bob Hanson