adds
 moveto [optional-time] PLACE

where PLACE is one of:

  front, back, left, right, top, bottom  

This to implement in script what is possible with the application.
Also added to the applet menu as "View->"

Also introduces for select:

atomIndex, like atomNo, but the internal Jmol index value for the atom. 

the XOR operator and "tog" operators

XOR I thought might be useful -- it flips bits ON/OFF

"TOG" is just my invention so that we have a script-equivalent of atom picking toggles. 

A tog B :: (A and B) == B ? A not B : A or B

That is, if all the bits of B are set in A, then clear them from A;
otherwise do a logical OR.

So a typical "set picking group" click looks like:

  select selected tog (within(group, atomindex=324))


Then