applet: adds support for using javascript evaluation in script commands in place of a filename:
 script "javascript:getTheScript()"

instead of

 script "myfile.spt"

thus solving the stubborn issue of the applet losing 
all line formatting when reading scripts and data at load time
via the <param> tags. 

All you do is give the same sort of call to javascript that 
you would on the browser location line:

jmolApplet(600,'script "javascript:getScript()"')

The function getScript() is responsible for delivering the 
actual script to be run, and it is run at the correct time, 
since the applet only executes this script command when it
is ready to load the data. And since a script can now include

data "model data"
...
end "model data";

sequences, this means that raw file data or state scripts can
be run automatically at load time. All one needs to do is put
the data to be loaded into any textarea on the page.

This opens the way to transmitting serialized state data 
embedded in email messages as attachments. 

Bob Hanson