11.1 -- new frame/model specification options
When specifying a model from a multimodel set, as in

display */3

the number 3 is the model number, which is not necessarily 
the third model in the file. (For example, if the PDB MODEL
line reads "51" for the first model, then "display */51" 
displays that model.) This is Jmol 10 behavior.

New here is the option to specify the model by its order in the
file, not its "model number". This is done using decimal notation:

  file.model

display */1.2

displays the second model in the first (possibly only) loaded file.

display */2.3

displays the third model in the second-loaded file.

Since there is no ambiguity, "*/" is optional:

load "fileset" "set1.xyz" "set2.pdb"
display 1.2, 1.3, 1.4, 2.3

displays the second, third, and fourth models in set1.xyz 
along with the third model in file set2.pdb.

FRAME/ANIMATION/MODEL COMMAND 
-----------------------------

The decimal notation also works in the frame command, 
aka "animation" or "model"

frame 1.2  #displays first file, second model
animation range 1.2 1.5 #sets the animation range to 1.2-1.5


INTERACTION OF DISPLAY AND FRAME COMMANDS
-----------------------------------------

If both display and frame commands are used, be aware that
they are independent. If you say:

model 1.2
display 1.3

You will see nothing, because you have set the model to 1.2
and are only displaying atoms in model 1.3. This is intentional.
When using display across models, it is best to do something
like:

models all
display 1.2

so that you are certain to get a display.

To display a selected set of atoms in a set of models, you 
should do something like the following:

model 0
display (1.2,1.3,1.4) and carbon

or

model 0
display (1.2,1.3,1.4) 
display displayed and carbon

The commas here are important.