11.1.14 expands operand type to include bitsets and uses better notation for bitset ranges:
x = {carbon}    # the number "2"
x = {carbon}[2]    # 2nd carbon in the set
x = {carbon}[2][4] # 2nd through 4th in the set
x = {carbon}[2][0] # all carbons after the first

Now "[]" is an operator, so you can do something like this:

x = {carbon}[{carbon}]  # the last carbon in the set.

The first {carbon} here is a bitset. The second is, too, but because
we need an integer in [ ], that one is automatically turned into its 
cardinality (the number of 1s in the bitset).

a = 3

x = {carbon}[a].xyz  # the xyz coordinate of the a-th carbon in the set.


Obviously, this could be nice for select as well, but we can't be
doing things like:

select {oxygen}[3][5]

so I've implemented the following:

set x = {carbon}[3][5]

then "x" = 3 (because there are 3 atoms, #3, #4, and #5)
and
"~x" represents the bitset for select:

select ~x

gives you those three atoms selected.

The reason for the tilde is to make sure there is no conflict with 
Jmol predefined names.