11.1.18 adds:
--capability to apply any atom-based user data from external files 
to color an isosurface or as text for atom labels.

--isosurface extension to APBS OpenDX output files, allowing 
very nice rendering of molecular electrostatic potentials. 
JVXL compression 300:1 is observed for the color mapping of a
protein surface. 

--addition of percent-based applet sizing in Jmol.js allows 
automatic resizing of the applet when the page is resized. All
you have to do is use a decimal such as 0.95 for the width:

jmolApplet(0.95,"...")

to indicate percent (95% width and 95% height in this case)

Highly recommended to use "set zoomLarge=false" with this so
that the model stays in the window on all resizing.

--testscripts JUnit testing (incomplete)