11.1.23 adds isosurface CAVITY x.xx # where x.xx is 1 to 10
Isosurface CAVITY is an invented concept implemented only in Jmol.
This keyword does nothing by itself other than set the surfaceDistance
of atoms. The surfaceDistance measure approximates the 
distance from an atom to the nearest point on a "skin" 
tightened around the model, thus excentuating molecular cavities. 

The adjustable parameter x.xx determines the "tightness" of 
the skin, where a very tight skin (10.0) would touch only the very
outermost atoms, and a loose skin (1.0) would touch every atom
near on the standard "molecular" surface ("sasurface 0"). 

The keyword can be used by itself:

isosurface select(*) ignore(solvent) cavity 10

so that later isosurface commands involving surfaceDistance
use these colors:

isosurface s1 select(*) ignore(solvent) cavity 10
isosurface within 10.0 (atomno=2345) map property solventDistance

or in that same command:

isosurface s1 select(*) ignore(solvent) cavity 10 sasurface 0 map property surfaceDistance

Note that currently the cavity calculation usurps dots; the
two are not compatible. This is because the calculation works 
as follows:

1) A dot surface is created extending out x.xx Angstroms from
the molecule. A large number such as 10.0 for x.xx effectively
smooths out "wrinkles" in the molecule (cavities), and exposes
only the very outermost atoms. The atom's "surfaceDistance"
amounts to the distance from the atom to the nearest sample point
minus x.xxx. Atoms in the interior or along the sides of a 
cavity less than two times x.xxx in diameter are thus distinguished
from the outer "skin" atoms, which have surfaceDistance = 0.
 
This calculation replaces the previous "surfaceDistance" 
calculation, which also usurped dots. That calculation was 
really "distance to the nearest surface atom" -- which did not
accurately represent distance to the nearest surface in the 
case of cavities. 
were not