Making function calls thread-safe.
ScriptEvaluator: Given function object holds state, it can not be called from two threads at the same time. So, runFunction is wrapped in synchronized(function).
ParallelProcessor: Is modified to use lock object vs this as this is used in ScriptEvaluator.
Viewer: Change to Executor factory to use different thread executor, thus avoiding deadlock if parallel functions call parallel functions and allows parallel processes to be eligible to run.